fix(client): Arrumado bugs com a nova feature do builder
- Instancias Builder ultilizadas no codigo para melhorar a legibilidade.

+++ src/main/java/com/agencia/project/service/ClientService.java
@@ -9,12 +9,12 @@ public final class ClientService {
 
     public static void save(ClientModel client) { ClientDAO.createClient(client); }
 
-    public static ClientModel findById(int id) {
+    public static ClientModel findById(long id) {
         Optional<ClientModel> clientModel = ClientDAO.readClientById(id);
         return clientModel.orElseThrow(() -> new NullPointerException("Cliente nao encontrado"));
     }
 
     public static void update(ClientModel client) { ClientDAO.updateClient(client); }
 
-    public static void delete(int id) { ClientDAO.deleteClient(id); }
+    public static void delete(long id) { ClientDAO.deleteClient(id); }
 }
